requirements.txt
openai>=1.0.0
pandas>=1.3.0
numpy>=1.20.0
matplotlib>=3.4.0
scikit-learn>=1.0.0
tqdm>=4.62.0
pathlib>=1.0.1
seaborn>=0.11.0
python-dotenv>=0.19.0 